feat: add contact event by service account (#149)
* feat: add postContactEventWithPlatineServiceAccount controller

* feat: add postContactEventWithPlatineServiceAccount controller

* chore: update pom.xml

* fix: update constant name for portal privilege endpoint

Renamed the constant for the contact events endpoint from "platine-account" to "portal-privilege" to better reflect its purpose. This enhances clarity and consistency within the API definition.

* fix: identifier case sensitivity issue in contact event check

Removed forced uppercase conversion when validating contact identifiers. This ensures the identifier is handled as provided, maintaining consistency and preventing unnecessary mismatches.

* fix : Remove unused convertToEntity method from ContactEventController

The convertToEntity method was not being used and has been removed to clean up the code. This improves maintainability and eliminates unnecessary clutter in the controller.